requirements.txt
asgiref==3.8.1
boto3==1.35.76
botocore==1.35.76
dj-database-url==2.3.0
Django==4.2.16
django-ckeditor==6.7.1
django-heroku==0.3.1
django-js-asset==2.2.0
django-storages==1.14.4
jmespath==1.0.1
pillow==11.0.0
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.2
typing_extensions==4.12.2
urllib3==1.26.20
whitenoise==6.8.2
